Added spawn eggs, changed agility loss to only happen when eating or sleeping, asphyxiation now causes organ damage and as a result cyanide now actually kills you, fixed mobs not being able to spawn

+++ src/main/java/com/site21/bittermelon/content/character/skills/Skill.java
@@ -16,9 +16,9 @@ public enum Skill implements StringRepresentable {
     BRAVERY("bravery", 2);
 
     private final String name;
-    private final int maxLevel;
+    private final float maxLevel;
 
-    Skill(String name, int maxLevel) {
+    Skill(String name, float maxLevel) {
         this.name = name;
         this.maxLevel = maxLevel;
     }
@@ -28,7 +28,7 @@ public enum Skill implements StringRepresentable {
         return name;
     }
 
-    public int getMaxLevel() {
+    public float getMaxLevel() {
         return maxLevel;
     }
